=Source *Fixed a type in skill.c for the updated renewal casting/cycling system.
=Other
In the last revision I wanted to thank Inkfish for making this code which limited the renewal cast times
to 3rd job skills and added the 2nd required dex setting for 3rd jobs to instant cast 1st and 2nd job skills.
Also I want to thank Kazumi for letting me add this code to 3CeAM since it was made as a custom for her server.
What I take credit for here is a small modification I made to the code where it now properly check if theirs a
fixed time and if theirs none to give no fixed. Mainly the removal of that 20% fixed stuff.
Thank you so much both of you for this. 3CeAM made a big leap because of this. =D

git-svn-id: http://3ceam.googlecode.com/svn/trunk@640 ca0df34e-21c1-4b9e-08a3-f23acb074bb3